Move `SignerStatusPerZone` to `signer/status.rs`
